T 29 migrate to net6

+++ T034.Core/Entity/Files.cs
@@ -11,7 +11,7 @@ namespace T034.Core.Entity
         public virtual Folder Folder { get; set; }
         
         public virtual DateTime LogDate { get; set; }
-        public virtual int Size { get; set; }
+        public virtual long Size { get; set; }
         public virtual User User { get; set; }
     }
 }
